requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
pygithub>=2.1.1
gitpython>=3.1.40
requests>=2.31.0
python-dotenv>=1.0.0

# Web server for JavaScript UI
flask>=3.0.0
flask-cors>=4.0.0

# Database for session storage
pymongo>=4.15.0

# AI SDK and dependencies (compatible with multiple providers)
openai>=1.6.1
tiktoken>=0.12.0
sniffio>=1.3.0
distro>=1.9.0
jiter>=0.10.0
tqdm>=4

# Note: These packages are required by langchain-openai for SDK compatibility
# They support multiple AI providers including Gemini, GPT, Claude, etc.
